Clean up the code and the layout
