+++ App/Settings/AwfulIcon.swift
@@ -1,0 +1,15 @@
+//
+//  AwfulIcon.swift
+//  Awful
+//
+//  Created by Liam Westby on 9/24/17.
+//  Copyright © 2017 Awful Contributors. All rights reserved.
+//
+
+import UIKit
+
+class AwfulIcon: UICollectionViewCell {
+    @IBOutlet weak var imageView: UIImageView!
+    
+
+}
