+++ App/Settings/AwfulIconCollection.swift
@@ -1,0 +1,30 @@
+//
+//  AwfulIconCellTableViewCell.swift
+//  Awful
+//
+//  Created by Liam Westby on 9/23/17.
+//  Copyright © 2017 Awful Contributors. All rights reserved.
+//
+
+import UIKit
+
+class AwfulIconCollection: UITableViewCell, UICollectionViewDataSource, UICollectionViewDelegate {
+    @IBOutlet weak var collection: UICollectionView?
+    var selectedIconName: String = "v"
+    
+    override func awakeFromNib() {
+        collection?.dataSource = self
+        collection?.register(UINib(nibName: "AwfulIcon", bundle: nil), forCellWithReuseIdentifier: "appIcon")
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 20
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let collection = collection else { Logger.get().e("No collection in view"); return UICollectionViewCell() }
+        let cell = collection.dequeueReusableCell(withReuseIdentifier: "appIcon", for: indexPath) as! AwfulIcon
+        cell.imageView.image = UIImage(named: "AppIcon-\(selectedIconName)-60x60", in: Bundle(for: type(of: self)), compatibleWith: nil)
+        return cell
+    }
+}
