+++ App/Settings/SettingsSliderCell.xib
@@ -1,7 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11201" systemVersion="16A323" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="13196" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
     <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11161"/>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13173"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
@@ -12,6 +16,7 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <slider opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" value="0.5" minValue="0.0" maxValue="1" minimumValueImage="brightness-down" maximumValueImage="brightness-up" translatesAutoresizingMaskIntoConstraints="NO" id="USL-U7-cGb">
+                    <rect key="frame" x="18" y="15" width="339" height="31"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="30" id="tIw-4N-sSx"/>
                     </constraints>
@@ -34,7 +39,7 @@
         </view>
     </objects>
     <resources>
-        <image name="brightness-down" width="25" height="25"/>
-        <image name="brightness-up" width="25" height="25"/>
+        <image name="brightness-down" width="30" height="30"/>
+        <image name="brightness-up" width="30" height="30"/>
     </resources>
 </document>
